Login Started

+++ BlankSpace/Controllers/BusScheduleController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using BlankSpace.Database;
 using BlankSpace.Models;
 using BlankSpace.ViewModels;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -30,14 +31,23 @@ namespace BlankSpace.Controllers
         }
         public IActionResult AddSchedule()
         {
-            var bus = _context.Buses.AsNoTracking().ToList();
-            ViewBag.Bus = new SelectList(bus, "BusId", "CoachName");
+            if (HttpContext.Session.GetString("UserRole") == "1")
+            {
+                var bus = _context.Buses.AsNoTracking().ToList();
+                ViewBag.Bus = new SelectList(bus, "BusId", "CoachName");
 
-            var start = _context.Places.AsNoTracking().ToList();
-            ViewBag.Start = new SelectList(start, "PlaceId", "PlaceName");
+                var start = _context.Places.AsNoTracking().ToList();
+                ViewBag.Start = new SelectList(start, "PlaceId", "PlaceName");
 
-           
-            return View();
+
+                return View();
+
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         } 
         [HttpPost]
         public IActionResult AddSchedule(BusScheduleVm a)
@@ -60,54 +70,65 @@ namespace BlankSpace.Controllers
         }
         public IActionResult ScheduleList()
         {
-            var ss = _context.BusSchedules.AsNoTracking().Include(sp=>sp.Bus).ToList();
-            var p = _context.Places.AsNoTracking().ToList();
-            var pp = _context.Places.AsNoTracking().ToList();
-            var s = from a in ss
-                    join ab in p on a.StartingFrom equals ab.PlaceId
-                    join aaab in pp on a.Destination equals aaab.PlaceId
-                  
-                    select new {
-
-                        BusScheduleId = a.BusScheduleId,
-                        DestinationName = aaab.PlaceName,
-                        StartingFromName = ab.PlaceName,
-                        Time = a.Time,
-                      
-                        CoachName = a.Bus.CoachName,
-                        TicketPrice = a.TicketPrice
-                    };
-
-
-
-            var sent = new List<BusScheduleVm>();
-            int c = 1;
-            foreach (var item in s)
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                BusScheduleVm sp = new BusScheduleVm()
-                {
-                    BusScheduleId = item.BusScheduleId,
-                    DestinationName = item.DestinationName,
-                    StartingFromName = item.StartingFromName,
-                    Time = item.Time,
-                    Serial = c,
-                    CoachName=item.CoachName,
-                    TicketPrice=item.TicketPrice
+                var ss = _context.BusSchedules.AsNoTracking().Include(sp => sp.Bus).ToList();
+                var p = _context.Places.AsNoTracking().ToList();
+                var pp = _context.Places.AsNoTracking().ToList();
+                var s = from a in ss
+                        join ab in p on a.StartingFrom equals ab.PlaceId
+                        join aaab in pp on a.Destination equals aaab.PlaceId
 
-                };
-                sent.Add(sp);
-                  c++;
+                        select new
+                        {
+
+                            BusScheduleId = a.BusScheduleId,
+                            DestinationName = aaab.PlaceName,
+                            StartingFromName = ab.PlaceName,
+                            Time = a.Time,
+
+                            CoachName = a.Bus.CoachName,
+                            TicketPrice = a.TicketPrice
+                        };
+
+
+
+                var sent = new List<BusScheduleVm>();
+                int c = 1;
+                foreach (var item in s)
+                {
+                    BusScheduleVm sp = new BusScheduleVm()
+                    {
+                        BusScheduleId = item.BusScheduleId,
+                        DestinationName = item.DestinationName,
+                        StartingFromName = item.StartingFromName,
+                        Time = item.Time,
+                        Serial = c,
+                        CoachName = item.CoachName,
+                        TicketPrice = item.TicketPrice
+
+                    };
+                    sent.Add(sp);
+                    c++;
+                }
+
+
+
+                return View(sent);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
             }
            
-           
-
-            return View(sent);
         } 
         public IActionResult DetailsSchedule(int id)
         {
-            var s = _context.BusSchedules.AsNoTracking().Where(sa => sa.BusScheduleId == id).Include(p=>p.Bus).FirstOrDefault();
-            var Destination = _context.Places.Where(ss => ss.PlaceId == s.Destination).FirstOrDefault();
-            var Start = _context.Places.Where(ss => ss.PlaceId == s.StartingFrom).FirstOrDefault();
+            if (HttpContext.Session.GetString("UserRole") == "1")
+            {
+                var s = _context.BusSchedules.AsNoTracking().Where(sa => sa.BusScheduleId == id).Include(p => p.Bus).FirstOrDefault();
+                var Destination = _context.Places.Where(ss => ss.PlaceId == s.Destination).FirstOrDefault();
+                var Start = _context.Places.Where(ss => ss.PlaceId == s.StartingFrom).FirstOrDefault();
 
 
 
@@ -115,33 +136,47 @@ namespace BlankSpace.Controllers
 
 
 
-            BusScheduleVm sp = new BusScheduleVm()
+                BusScheduleVm sp = new BusScheduleVm()
                 {
                     BusScheduleId = s.BusScheduleId,
                     DestinationName = Destination.PlaceName,
                     StartingFromName = Start.PlaceName,
                     Time = s.Time,
-                    TicketPrice=s.TicketPrice,
-                    CoachName=s.Bus.CoachName
+                    TicketPrice = s.TicketPrice,
+                    CoachName = s.Bus.CoachName
                 };
-            return View(sp);
+                return View(sp);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
         public IActionResult DeleteSchedule(int id)
         {
-            var s = _context.BusSchedules.AsNoTracking().Where(sa => sa.BusScheduleId == id).Include(a=>a.Bus).FirstOrDefault();
-            var Destination = _context.Places.Where(ss => ss.PlaceId == s.Destination).FirstOrDefault();
-            var Start = _context.Places.Where(ss => ss.PlaceId == s.StartingFrom).FirstOrDefault();
-
-            BusScheduleVm sp = new BusScheduleVm()
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                BusScheduleId = s.BusScheduleId,
-                DestinationName = Destination.PlaceName,
-                StartingFromName = Start.PlaceName,
-                Time = s.Time,
-                TicketPrice = s.TicketPrice,
-                CoachName = s.Bus.CoachName
-            };
-            return View(sp); 
+                var s = _context.BusSchedules.AsNoTracking().Where(sa => sa.BusScheduleId == id).Include(a => a.Bus).FirstOrDefault();
+                var Destination = _context.Places.Where(ss => ss.PlaceId == s.Destination).FirstOrDefault();
+                var Start = _context.Places.Where(ss => ss.PlaceId == s.StartingFrom).FirstOrDefault();
+
+                BusScheduleVm sp = new BusScheduleVm()
+                {
+                    BusScheduleId = s.BusScheduleId,
+                    DestinationName = Destination.PlaceName,
+                    StartingFromName = Start.PlaceName,
+                    Time = s.Time,
+                    TicketPrice = s.TicketPrice,
+                    CoachName = s.Bus.CoachName
+                };
+                return View(sp);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
         [HttpPost]
         public IActionResult DeleteSchedule(BusScheduleVm a)
@@ -154,46 +189,69 @@ namespace BlankSpace.Controllers
         } 
         public IActionResult UpdateSchedule(int  id) 
         {
-            var bus = _context.Buses.AsNoTracking().ToList();
-            ViewBag.Bus = new SelectList(bus, "BusId", "CoachName");
-            var Route1 = _context.Places.AsNoTracking().ToList();
-            ViewBag.Road = new SelectList(Route1, "PlaceId", "PlaceName");
-            var s = _context.BusSchedules.AsNoTracking().Where(sa => sa.BusScheduleId == id).FirstOrDefault();
-
-            BusScheduleVm sp = new BusScheduleVm()
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                BusScheduleId = s.BusScheduleId,
-                Destination = s.Destination,
-                StartingFrom = s.StartingFrom,
-                Time = s.Time,
-                BusId=s.BusId,
-                TicketPrice=s.TicketPrice
-            };
-            return View(sp);
+                var bus = _context.Buses.AsNoTracking().ToList();
+                ViewBag.Bus = new SelectList(bus, "BusId", "CoachName");
+                var Route1 = _context.Places.AsNoTracking().ToList();
+                ViewBag.Road = new SelectList(Route1, "PlaceId", "PlaceName");
+                var s = _context.BusSchedules.AsNoTracking().Where(sa => sa.BusScheduleId == id).FirstOrDefault();
+
+                BusScheduleVm sp = new BusScheduleVm()
+                {
+                    BusScheduleId = s.BusScheduleId,
+                    Destination = s.Destination,
+                    StartingFrom = s.StartingFrom,
+                    Time = s.Time,
+                    BusId = s.BusId,
+                    TicketPrice = s.TicketPrice
+                };
+                return View(sp);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
         [HttpPost]
         public IActionResult UpdateSchedule(BusScheduleVm a)
         {
-            BusSchedule s = new BusSchedule()
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                BusScheduleId = a.BusScheduleId,
-                Destination = a.Destination,
-                StartingFrom = a.StartingFrom,
-                Time = a.Time,
-                BusId = a.BusId,
-                TicketPrice = a.TicketPrice
+                BusSchedule s = new BusSchedule()
+                {
+                    BusScheduleId = a.BusScheduleId,
+                    Destination = a.Destination,
+                    StartingFrom = a.StartingFrom,
+                    Time = a.Time,
+                    BusId = a.BusId,
+                    TicketPrice = a.TicketPrice
 
 
-            };
-            _context.BusSchedules.Update(s);
-            _context.SaveChanges();
-            return RedirectToAction("ScheduleList");
+                };
+                _context.BusSchedules.Update(s);
+                _context.SaveChanges();
+                return RedirectToAction("ScheduleList");
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
 
         public IActionResult AddRoute()
         {
-
-            return View();
+            if (HttpContext.Session.GetString("UserRole") == "1")
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+           
         }
         [HttpPost]
         public IActionResult AddRoute(PlaceVm a)
@@ -210,16 +268,24 @@ namespace BlankSpace.Controllers
         }
         public IActionResult UpdateRoute(int id) 
         {
-            var item = _context.Places.AsNoTracking().Where(s => s.PlaceId == id).FirstOrDefault();
-
-            PlaceVm p = new PlaceVm()
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                PlaceName = item.PlaceName,
-                PlaceVmId = item.PlaceId,
+                var item = _context.Places.AsNoTracking().Where(s => s.PlaceId == id).FirstOrDefault();
 
-            };
+                PlaceVm p = new PlaceVm()
+                {
+                    PlaceName = item.PlaceName,
+                    PlaceVmId = item.PlaceId,
 
-            return View(p);
+                };
+
+                return View(p);
+            }
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            
         }
         [HttpPost]
         public IActionResult UpdateRoute(PlaceVm a)
@@ -238,43 +304,60 @@ namespace BlankSpace.Controllers
         }
         public IActionResult DeleteRoute(int id) 
         {
-            try
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                var sp = _context.Places.Where(s => s.PlaceId == id).FirstOrDefault();
-                _context.Places.Remove(sp);
-                _context.SaveChanges();
-                return RedirectToAction("RouteList");
+                try
+                {
+                    var sp = _context.Places.Where(s => s.PlaceId == id).FirstOrDefault();
+                    _context.Places.Remove(sp);
+                    _context.SaveChanges();
+                    return RedirectToAction("RouteList");
+
+                }
+                catch (Exception)
+                {
+                    ViewBag.Error = "You can not delete this as it is used in another table ";
+                    return RedirectToAction("RouteList");
+                }
 
             }
-            catch (Exception )
+            else
             {
-                ViewBag.Error = "You can not delete this as it is used in another table ";
-                return RedirectToAction("RouteList");
+                return RedirectToAction("Index", "Home");
             }
+           
 
 
             
         }
         public IActionResult RouteList()
         {
-            var a = _context.Places.AsNoTracking().ToList();
-            var s = new List<PlaceVm>();
-            int c = 1;
-            foreach (var item in a)
+            if (HttpContext.Session.GetString("UserRole") == "1")
             {
-                PlaceVm p = new PlaceVm()
+
+                var a = _context.Places.AsNoTracking().ToList();
+                var s = new List<PlaceVm>();
+                int c = 1;
+                foreach (var item in a)
                 {
-                    PlaceName = item.PlaceName,
-                    PlaceVmId = item.PlaceId,
-                    Serial=c
+                    PlaceVm p = new PlaceVm()
+                    {
+                        PlaceName = item.PlaceName,
+                        PlaceVmId = item.PlaceId,
+                        Serial = c
 
-                };
-                s.Add(p);
-                c++;
+                    };
+                    s.Add(p);
+                    c++;
 
+                }
+
+                return View(s);
             }
-            
-            return View(s);
+            else
+            {
+                return RedirectToAction("Index", "Home");
+            }
         }
 
     }
